requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
ConfigArgParse==1.5.3
contourpy==1.0.6
cycler==0.11.0
dash==2.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
fastjsonschema==2.16.2
filelock==3.8.2
Flask==2.2.2
fonttools==4.38.0
huggingface-hub==0.11.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_core==5.1.1
kiwisolver==1.4.4
logger==1.4
MarkupSafe==2.1.1
matplotlib==3.6.2
nbformat==5.5.0
numpy==1.24.0
open3d==0.16.1
packaging==22.0
Pillow==9.3.0
platformdirs==2.6.0
plotly==5.11.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
six==1.16.0
tenacity==8.1.0
tokenizers==0.13.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.8.0
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2

# package location
--find-links https://download.pytorch.org/whl/torch_stable.html